requirements.txt
afwf==0.0.3
atomicwrites==1.4.0
attrs==21.4.0
attrs-mate==1.0.2
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Deprecated==1.2.13
diskcache==5.4.0
fuzzywuzzy==0.18.0
idna==3.3
pathlib-mate==1.0.3
pycodestyle==2.8.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-Levenshtein==0.12.2
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.9
whoosh==2.7.4
wrapt==1.14.0
